Reorganize folder structure and code.
