+++ index.html
@@ -8,8 +8,7 @@
 			<!-- Google font used -->
 			<link rel="stylesheet" href="https://maxcdn.bootstrapcdn.com/font-awesome/4.3.0/css/font-awesome.min.css" />
 			<!-- fontawesome icons-->
-			<link rel="stylesheet" type="text/css" href="index.css" />
-			<meta name="image" property="og:image" content="/images/deezer_music_player_social_media_preview.jpg" />
+			<link rel="stylesheet" type="text/css" href="css/index.css" />
 		</head>
 		<body>
 			<div class="formsWrapper">
@@ -71,7 +70,7 @@
 			</div>
 			<div id="searchResultsContainer"></div>
 
-			<script type="module" src="index.js"></script>
+			<script type="module" src="js/index.js"></script>
 		</body>
 	</html>
