+++ package.json
@@ -1,0 +1,7 @@
+{
+	"devDependencies": {
+		"@eslint/js": "^9.39.2",
+		"eslint": "^9.39.2",
+		"globals": "^16.5.0"
+	}
+}
